Updates to CDK install instructions
The lab code and all the import constructs are for AWS CDK V1, however Cloud9 now comes with CDK V2 and the lab does not work if someone was to copy/paste all the instructions. It requires users to troubleshoot the error.

I've done a small update on the CDK install instructions where it removes the CDK V2 from the default Cloud 9 instance and installs CDK V1 where all the dependencies work perfectly. Also added a CDK alias so no additional workshop instructions needs to be updated.